% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalGLM.R
\name{causalGLM}
\alias{causalGLM}
\title{causalGLM
A default and more user-friendly front-end of the implemented methods.}
\usage{
causalGLM(
  formula,
  W,
  A,
  Y,
  estimand = c("CATE", "OR", "RR"),
  learning_method = c("autoHAL", "glm", "glmnet", "gam", "mars", "xgboost"),
  inference_type = c("semiparametric", "parametric"),
  cross_fit = TRUE,
  sl3_Learner = NULL,
  glm_formula_A = NULL,
  glm_formula_Y = NULL,
  glm_formula_Y0W = glm_formula_Y,
  weights = NULL,
  data_list = NULL,
  fast_analysis = TRUE,
  parallel = F,
  ncores = 4,
  smoothness_order = 1,
  max_degree = 2,
  num_knots = c(10, 5)
)
}
\arguments{
\item{formula}{A R formula object specifying the parametric form of CATE, OR, or RR (depending on method).}

\item{W}{A named matrix or data.frame of baseline covariates to condition on.}

\item{A}{A binary treatment assignment vector}

\item{Y}{n outcome variable (continuous, nonnegative or binary depending on method)}

\item{learning_method}{Machine-learning method to use. This is overrided if argument \code{sl3_Learner} is provided. Options are:
"autoHAL": Adaptive robust automatic machine-learning using the Highly Adaptive Lasso \code{hal9001}
"glm": Fit nuisances with parametric model. See arguments \code{glm_formula_A}, \code{glm_formula_Y} and \code{glm_formula_Y0}.
"glmnet": Learn using lasso with glmnet.
"gam": Learn using generalized additive models with mgcv.
"xgboost": Learn using a default cross-validation tuned xgboost library with max_depths 3 to 7.}

\item{inference_type}{What inferential method to use. Currently only supports semiparametric `sp` methods.}

\item{sl3_Learner}{A \code{sl3} Learner object to use to estimate nuisance functions with machine-learning.}

\item{glm_formula_A}{A glm formula for P(A=1|W). Only used if learning_method = "glm".}

\item{glm_formula_Y}{A glm formula for E[Y|A,W]. Only used if learning_method = "glm".}

\item{glm_formula_Y0W}{A glm formula for E[Y|A=0,W]. Only used if learning_method = "glm".}

\item{weights}{An optional vector of weights to use in procedure.}

\item{parallel}{Whether to parallize HAL whenever it is used in fitting.}

\item{ncores}{Number of cores to use in parallelization.}

\item{smoothness_order}{Smoothness order for HAL (see \code{hal9001/fit_hal})}

\item{max_degree}{Max interaction degree for HAL (see \code{hal9001/fit_hal})}

\item{num_knots}{Number of knots by interaction degree for HAL (see \code{hal9001/fit_hal}). Used to generate basis functions.}

\item{Estimand}{Estimand/parameter to estimate. Choices are:
CATE: Estimate conditional average treatment effect with \code{spCATE} assuming it satisfies parametric model \code{formula}.
OR: Estimate conditional odds ratio with \code{spOR} assuming it satisfies parametric model \code{formula}.
OR: Estimate conditional relative risk with \code{spRR} assuming it satisfies parametric model \code{formula}.}
}
\description{
causalGLM
A default and more user-friendly front-end of the implemented methods.
}
