% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.CATE}
\alias{sim.CATE}
\title{Simulate a dataset with known constant CATE}
\usage{
sim.CATE(n = 1500, p = 2)
}
\arguments{
\item{n}{Sample size}

\item{p}{Dimension of W}
}
\description{
Simulate a dataset with known constant CATE
}
