% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.OR}
\alias{sim.OR}
\title{Simulate a dataset with known constant OR}
\usage{
sim.OR(n = 1500, p = 2)
}
\arguments{
\item{n}{Sample size}

\item{p}{Dimension of W}
}
\description{
Simulate a dataset with known constant OR
}
